Conflicts:
	dataBaseManagament.sh
	startEngine.sh

+++ dataBaseManagament.sh
@@ -18,7 +18,9 @@ function listDatabases()
         echo "Empty"
         continue
       else
-        echo $d
+
+         IFS='/' read -ra my_array <<< "$d"
+         echo ${my_array[0]}
       
     fi
     done
@@ -27,19 +29,31 @@ function listDatabases()
 
 function connectToDatabase()
 {
-    databaseName=$1
+   databaseName=$1
+   databaseNameLen=`echo -n $databaseName | wc -m`
+   if [[ $databaseNameLen != 0 ]]
+   then
+      if [ -d "$databaseName" ];
+      then
+            #echo "1";
+            cd $databaseName
+            echo "-------------------------------"
+            echo -e "${Green}  connected to $databaseName db ${NC}"
+            echo "-------------------------------"
+            . $startLocation/dataTableManagament.sh $startLocation $databaseName
 
-    if [ -d "$databaseName" ];
-    then
-         #echo "1";
-         cd $databaseName
-         echo "-------------------------------"
-         echo -e "${Green}  connected to $databaseName db ${NC}"
-         echo "-------------------------------"
-         #ls -l 
-         #sleep 5
-         #echo "here now $PWD" 
+            #ls -l 
+            #sleep 5
+            #echo "here now $PWD" 
 
+      else
+         #echo "0";
+         
+            echo "-----------------------------"
+            echo -e "${Yellow} $databaseName db does not exist ${NC}"
+            echo "-----------------------------"
+      fi
+                 
     else
       #echo "0";
       
@@ -51,40 +65,61 @@ function connectToDatabase()
 
 function createDatabase()
 {
-    databaseName=$1
 
-    if [ -d "$databaseName" ];
-    then
-      #echo "1";
-         echo "---------------------------"
-         echo -e "${Yellow}the database already exists ${NC}"
-         echo "---------------------------"
+   databaseName=$1
+   databaseNameLen=`echo -n $databaseName | wc -m`
+   if [[ $databaseNameLen != 0 ]]
+   then
+      if [ -d "$databaseName" ];
+      then
+         #echo "1";
+            echo "---------------------------"
+           echo -e "${Yellow}the database already exists ${NC}"
+            echo "---------------------------"
 
+      else
+         #echo "0";
+            mkdir $databaseName
+            echo "---------------------------"
+            echo -e "${Green} $databaseName db has been created ${NC}"
+            echo "---------------------------"
+      fi
+                 
     else
-      #echo "0";
-         mkdir $databaseName
-         echo "---------------------------"
-         echo -e "${Green} $databaseName db has been created ${NC}"
-         echo "---------------------------"
-    fi
+    echo "-------------------------------------"
+     echo -e "${Red}Please Check your synax and try again ${NC}"
+    echo "-------------------------------------"
+   fi
+ 
 }
 function dropDatabase()
 {
-    databaseName=$1
-    if [ -d "$databaseName" ];
-    then
-      #echo "1";
-        rm -r $databaseName;
-         echo "------------------------"
-         echo -e "${Green} $databaseName has been deleted ${NC}"
-         echo "------------------------"
+    
+   databaseName=$1
+   databaseNameLen=`echo -n $databaseName | wc -m`
+   if [[ $databaseNameLen != 0 ]]
+   then
+         databaseName=$1
+      if [ -d "$databaseName" ];
+      then
+         #echo "1";
+         rm -r $databaseName;
+            echo "------------------------"
+            echo -e "${Green}$databaseName has been deleted ${NC}"
+            echo "------------------------"
 
+      else
+         #echo "0";
+            echo "------------------------------"
+              echo -e "${Yellow} $databaseName db does not exist ${NC}"
+            echo "------------------------------"
+      fi
+                 
     else
-      #echo "0";
-         echo "------------------------------"
-         echo -e "${Yellow} $databaseName db doesn't exist ${NC}"
-         echo "------------------------------"
-    fi
+    echo "-------------------------------------"
+   echo -e "${Red}Please Check your synax and try again ${NC}"
+    echo "-------------------------------------"
+   fi
 
 }
